feat: trainer profile page with data from api
